+++ src/db/migrations/20190408115030-add-validation-columns-users.js
@@ -1,49 +1,25 @@
 module.exports = {
-    up(queryInterface, Sequelize) {
-        /*
-      Add altering commands here.
-      Return a promise to correctly handle asynchronicity.
-      */
-        return queryInterface.addColumn('Users', 'contactMethodVerificationToken', {
+    up: async (queryInterface, Sequelize) => {
+        await queryInterface.addColumn('Users', 'contactMethodVerificationToken', {
             type: Sequelize.STRING,
             unique: true,
-        }).then( function () {
-            return queryInterface.addColumn('Users', 'contactMethodVerificationTokenExpires', {
-                type: Sequelize.DATE,
-            })
-        }).then( function () {
-            return queryInterface.addColumn('Users', 'contactMethodToVerify', {
-                type: Sequelize.STRING, // eslint-disable-line new-cap
-                defaultValue: '',
-            })
-        }).then( function () {
-            return queryInterface.addColumn('Users', 'verifiedContactMethods', {
-                type: Sequelize.ARRAY(Sequelize.STRING), // eslint-disable-line new-cap
-                defaultValue: [],
-            })
-        })
+        });
+        await queryInterface.addColumn('Users', 'contactMethodVerificationTokenExpires', {
+            type: Sequelize.DATE,
+        });
+        await queryInterface.addColumn('Users', 'contactMethodToVerify', {
+            type: Sequelize.STRING, // eslint-disable-line new-cap
+            defaultValue: '',
+        });
+        await queryInterface.addColumn('Users', 'verifiedContactMethods', {
+            type: Sequelize.ARRAY(Sequelize.STRING), // eslint-disable-line new-cap
+            defaultValue: [],
+        });
     },
-    down(queryInterface) {
-        /*
-      TODO: JRB THIS IS NOT RIGHT AND I KNOW IT. Or is it @Ryan?
-      */
-        return queryInterface.removeColumn('Users', 'contactMethodVerificationToken', {
-            type: Sequelize.STRING,
-            unique: true,
-        }).then( function () {
-            return queryInterface.removeColumn('Users', 'contactMethodVerificationTokenExpires', {
-                type: Sequelize.DATE,
-            })
-        }).then( function () {
-            return queryInterface.removeColumn('Users', 'contactMethodToVerify', {
-                type: Sequelize.STRING, // eslint-disable-line new-cap
-                defaultValue: '',
-            })
-        }).then( function () {
-            return queryInterface.removeColumn('Users', 'verifiedContactMethods', {
-                type: Sequelize.ARRAY(Sequelize.STRING), // eslint-disable-line new-cap
-                defaultValue: [''],
-            })
-        })
-    }
+    down: async (queryInterface) => {
+        await queryInterface.removeColumn('Users', 'contactMethodVerificationToken');
+        await queryInterface.removeColumn('Users', 'contactMethodVerificationTokenExpires');
+        await queryInterface.removeColumn('Users', 'contactMethodToVerify');
+        await queryInterface.removeColumn('Users', 'verifiedContactMethods');
+    },
 };